LayoutTests/scrollingcoordinator/mac/latching/horizontal-overflow-in-vertical-overflow-expected.txt

Scrolls vertically
Scrolls horizontally
Initial state
PASS outerScroller.scrollTop is 0
PASS innerScroller.scrollTop is 0
After scrolll
PASS outerScroller.scrollTop is 210
PASS innerScroller.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
